requirements.txt
feedparser==6.0.8
requests==2.31.0
Flask==2.2.2
slack_sdk==3.20.2
discord.py==2.0.1
matrix-nio==0.19.0
python-telegram-bot==22.3
PySocks==1.7.1

